# openapi.yaml
openapi: 3.0.3
info:
  title: Steroid Chat API
  description: |
    A real-time chat application where users join rooms by code.
    Supports text, images, files, and links. Includes ban system for moderation.
  version: 1.0.0
  contact:
    name: Steroid Support

servers:
  - url: http://localhost:3000/api
    description: Local development server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Users
    description: User management
  - name: Rooms
    description: Chat room management
  - name: Messages
    description: Message operations
  - name: Admin
    description: Administrative actions (admin only)

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      description: Create a new Steroid account. Username and email must be unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: cooluser
              email: user@example.com
              password: secret123
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      tags: [Auth]
      summary: Log into an existing account
      description: Returns a JWT token valid for 7 days.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account is banned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /rooms:
    get:
      tags: [Rooms]
      summary: List rooms the current user is a member of
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Rooms]
      summary: Create a new chat room
      description: Creates a room and returns a unique 6-character code to share.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /rooms/{code}:
    get:
      tags: [Rooms]
      summary: Get room details by code
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/NotFound'

  /rooms/{code}/join:
    post:
      tags: [Rooms]
      summary: Join a room by its code
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '200':
          description: Successfully joined room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/NotFound'

  /rooms/{code}/messages:
    get:
      tags: [Messages]
      summary: Get messages in a room
      description: Returns up to 50 messages in ascending order (oldest first).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
        - name: before
          in: query
          description: Fetch messages before this message ID (for pagination)
          schema:
            type: integer
        - name: limit
          in: query
          description: Number of messages to return (max 100)
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Messages]
      summary: Send a message to a room
      description: |
        Send text, image, file, or link. Use multipart/form-data.
        - For text: set `content` and `type=text`
        - For image/file: attach `file` field
        - For link: set `type=link` and `link_url`
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /admin/users/{userId}/ban:
    post:
      tags: [Admin]
      summary: Ban a user
      description: Bans the user from logging in or sending messages.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BanRequest'
      responses:
        '200':
          description: User banned
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Admin]
      summary: Unban a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User unbanned
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    RoomCode:
      name: code
      in: path
      required: true
      description: 6-character room code
      schema:
        type: string
        pattern: '^[A-Z0-9]{6}$'

    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer

  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT bearer token (expires in 7 days)
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        is_admin:
          type: boolean
        is_banned:
          type: boolean
        ban_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    Room:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: AB1C2D
        name:
          type: string
        created_by:
          type: integer
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time

    Message:
      type: object
      properties:
        id:
          type: integer
        room_id:
          type: integer
        user_id:
          type: integer
        username:
          type: string
        content:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, file, link]
        file_url:
          type: string
          nullable: true
        file_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    CreateRoomRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100

    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
          description: Text content (required for text/link types)
        type:
          type: string
          enum: [text, image, file, link]
          default: text
        file:
          type: string
          format: binary
          description: File to upload (image or any file)
        link_url:
          type: string
          format: uri
          description: URL for link type messages

    BanRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

  responses:
    BadRequest:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions or account banned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'